프론트 모듈에서 회원 등록,수정/ 로그인 로직 추가

+++ domain/src/test/java/sion/bookstore/domain/member/repository/MemberRepositoryTest.java
@@ -24,7 +24,7 @@ public class MemberRepositoryTest {
     public void findOneById() {
         Member expected = MemberMock.getMember("test", "test1@naver.com");
         memberRepository.create(expected);
-        Integer id = expected.getId();
+        Long id = expected.getId();
 
         Member actual = memberRepository.findOneById(id);
